All sub menus under option 1

+++ ObririUssd/Migrations/UssdDataContextModelSnapshot.cs
@@ -21,9 +21,9 @@ namespace ObririUssd.Migrations
 
             modelBuilder.Entity("ObririUssd.Models.UssdTransaction", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int")
+                        .HasColumnType("bigint")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<float>("Amount")
